+++ src/avt/Expressions/General/avtNormalizeExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                             avtNormalizeExpression.C                           //
+//                        avtNormalizeExpression.C                           //
 // ************************************************************************* //
 
 #include <avtNormalizeExpression.h>
@@ -143,13 +143,13 @@ avtNormalizeExpression::DeriveVariable(vtkDataSet *in)
 
     for (int i=0; i<ntuples; i++)
     {
-        float xin = vectorValues->GetComponent(i, 0);
-        float yin = vectorValues->GetComponent(i, 1);
-        float zin = vectorValues->GetComponent(i, 2);
+        double xin = vectorValues->GetComponent(i, 0);
+        double yin = vectorValues->GetComponent(i, 1);
+        double zin = vectorValues->GetComponent(i, 2);
 
-        float mag = sqrt(xin*xin + yin*yin + zin*zin);
+        double mag = sqrt(xin*xin + yin*yin + zin*zin);
 
-        float xout, yout, zout;
+        double xout, yout, zout;
         if (mag == 0.)
             xout = yout = zout = 0.;
         else
