+++ src/avt/Expressions/Math/avtCosExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                               avtCosExpression.C                              //
+//                               avtCosExpression.C                          //
 // ************************************************************************* //
 
 #include <avtCosExpression.h>
@@ -113,7 +113,7 @@ avtCosExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, cos(val));
         }
     }
