+++ src/avt/Expressions/Math/avtSquareRootExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                             avtSquareRootExpression.C                         //
+//                             avtSquareRootExpression.C                     //
 // ************************************************************************* //
 
 #include <avtSquareRootExpression.h>
@@ -107,13 +107,13 @@ avtSquareRootExpression::~avtSquareRootExpression()
  
 void
 avtSquareRootExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
-                                 int ncomponents, int ntuples)
+                                     int ncomponents, int ntuples)
 {
     for (int i = 0 ; i < ntuples ; i++)
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, sqrt(val));
         }
     }
