+++ src/avt/Expressions/Math/avtTanExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                               avtTanExpression.C                              //
+//                               avtTanExpression.C                          //
 // ************************************************************************* //
 
 #include <avtTanExpression.h>
@@ -107,13 +107,13 @@ avtTanExpression::~avtTanExpression()
  
 void
 avtTanExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
-                          int ncomponents, int ntuples)
+                              int ncomponents, int ntuples)
 {
     for (int i = 0 ; i < ntuples ; i++)
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, tan(val));
         }
     }
